src/requirements/base.txt
# Django
asgiref==3.8.1
Django==5.0.8
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.6.0

# Black
black==24.4.2
click==8.1.7
colorama==0.4.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2

# CORS
django-cors-headers==4.3.1

# Admin
django-admin-rangefilter==0.13.1

# Predict
tensorflow-cpu==2.17.0
scikit-learn==1.3.2
numpy==1.26.4
